used connection.row_factory = sqlite3.Row to access database as dictionary instead of default tuples, indices.
add_entry function used (?, ?) to avoid SQL injection attacks then
then add (,) and after that a tuple of data to change the ?.

get_entries create a cursor to retrieve data and also added
a comment to show alternative on how to retrieve data. Longer
but clear but we prefer it short.